webWhys/requirements.txt
# Website Competitor Scanner - Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# HTTP Client
aiohttp>=3.9.0

# HTML/Content Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0
readability-lxml>=0.8.0
tldextract>=5.0.0

# LLM Integration
litellm>=1.0.0
openai>=1.0.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
